Refactor SettingsScreen with modular components for improved code organization
